move tests to ray/train/tests folder
